Revert back to v1beta7
