+++ integration/examples/kaniko-local/skaffold.yaml
@@ -1,4 +1,4 @@
-apiVersion: skaffold/v1beta8
+apiVersion: skaffold/v1beta7
 kind: Config
 build:
   artifacts:
