+++ pkg/skaffold/schema/v1beta6/upgrade.go
@@ -17,8 +17,8 @@ limitations under the License.
 package v1beta6
 
 import (
+	next "github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/util"
-	next "github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/v1beta7"
 	pkgutil "github.com/GoogleContainerTools/skaffold/pkg/skaffold/util"
 	"github.com/pkg/errors"
 )
@@ -42,10 +42,21 @@ func (config *SkaffoldPipeline) Upgrade() (util.VersionedConfig, error) {
 			return nil, errors.Wrap(err, "converting new profile")
 		}
 	}
-	// convert Build (should be the same)
+
+	// Update profile if kaniko build exists
+	for i, p := range config.Profiles {
+		if err := upgradeKanikoBuild(p.Build, &newProfiles[i].Build); err != nil {
+			return nil, errors.Wrap(err, "upgrading kaniko build")
+		}
+	}
+
+	// convert Kaniko if needed
 	var newBuild next.BuildConfig
 	if err := pkgutil.CloneThroughJSON(config.Build, &newBuild); err != nil {
 		return nil, errors.Wrap(err, "converting new build")
+	}
+	if err := upgradeKanikoBuild(config.Build, &newBuild); err != nil {
+		return nil, errors.Wrap(err, "upgrading kaniko build")
 	}
 
 	// convert Test (should be the same)
@@ -62,4 +73,22 @@ func (config *SkaffoldPipeline) Upgrade() (util.VersionedConfig, error) {
 		Deploy:     newDeploy,
 		Profiles:   newProfiles,
 	}, nil
+}
+
+func upgradeKanikoBuild(build BuildConfig, newConfig *next.BuildConfig) error {
+	if build.KanikoBuild == nil {
+		return nil
+	}
+	kaniko := build.KanikoBuild
+	// Else, transition values from old config to new config artifacts
+	for _, a := range newConfig.Artifacts {
+		if err := pkgutil.CloneThroughJSON(kaniko, &a.KanikoArtifact); err != nil {
+			return errors.Wrap(err, "cloning kaniko artifact")
+		}
+	}
+	// Transition values from old config to in cluster details
+	if err := pkgutil.CloneThroughJSON(kaniko, &newConfig.Cluster); err != nil {
+		return errors.Wrap(err, "cloning cluster details")
+	}
+	return nil
 }
